[go] Move GtpClient to a new module
